StudyPal/requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-multipart==0.0.20
langchain==0.3.25
langchain-community==0.3.24
chromadb==1.0.12
pypdf==5.5.0
unstructured==0.17.2
openai==1.82.1
python-dotenv==1.1.0
python-jose==3.5.0
passlib==1.7.4
pydantic==2.11.5
pydantic-settings==2.9.1
typing-extensions==4.13.2
sqlalchemy==2.0.41
streamlit==1.45.1
requests==2.32.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyPDF2==3.0.1
docx==0.2.4
streamlit-option-menu==0.4.0